Added Crow2 fbx armature, mesh, texture, and ANIMATIONS
she should randomly and constantly be doing one of these, though they are overridden by any Use/Take actions the player does, and overridden by moving around.
"Crowidle1-100" plays 91% of the time
"Crowidle2-100" plays 3% of the time
"Crowidle3-200" plays 3% of the time (NOTE: ArmatureTome1 / Tome1 are a companion to this animation. she should be opening up a book and reading it, then putting it away. if this does not display properly, simply discard this animation in favor of Crowidle1-100 and Crowidle2-100)
"Crowidle4-200" plays 3% of the time (NOTE: ArmaturePotion1 / Potion1 are a companion to this animation. she should be taking out a suspicious bottle and drinking it, then putting it away. if this does not display properly, simply discard this animation in favor of Crowidle1-100 and Crowidle2-100)

Her walk animation is "Crowwalk1-60"

any time an item is PICKED UP, the animation "Crowuse1-60" should play when she reaches an acceptable destination (it does not need to actually look like she's picking something up)

When the game starts up, the animation "Crowwakeup-200" plays, and the screen fades from being black as it does so. She should be resting against the first car's left wall. This will be accompanied by some dialogue as she examines her situation.

She solves the final puzzle in the engine car with the animation "Crowmagicspell1-360", but you don't have to worry about this yet