redid sounds and added teafog again!
Check to see if "Rolling" is the new slidingdoor sound... if it is, the sounds should all have updated correctly.

Also let me know if Teafog reexported with more than 1 or 2 animations. I might have solved it but i'd like to know if not, so i can look into other solutions